+++ src/mms/iso_mms/common/mms_type_spec.c
@@ -3,181 +3,181 @@
  *
  *  Copyright 2013 Michael Zillgith
  *
- *	This file is part of libIEC61850.
+ *  This file is part of libIEC61850.
  *
- *	libIEC61850 is free software: you can redistribute it and/or modify
- *	it under the terms of the GNU General Public License as published by
- *	the Free Software Foundation, either version 3 of the License, or
- *	(at your option) any later version.
+ *  libIEC61850 is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
  *
- *	libIEC61850 is distributed in the hope that it will be useful,
- *	but WITHOUT ANY WARRANTY; without even the implied warranty of
- *	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *	GNU General Public License for more details.
+ *  libIEC61850 is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
  *
- *	You should have received a copy of the GNU General Public License
- *	along with libIEC61850.  If not, see <http://www.gnu.org/licenses/>.
+ *  You should have received a copy of the GNU General Public License
+ *  along with libIEC61850.  If not, see <http://www.gnu.org/licenses/>.
  *
- *	See COPYING file for the complete license text.
+ *  See COPYING file for the complete license text.
  */
 
 #include "libiec61850_platform_includes.h"
 #include "mms_common.h"
 #include "mms_type_spec.h"
+#include "mms_value_internal.h"
 
 void
 MmsVariableSpecification_destroy(MmsVariableSpecification* typeSpec)
 {
-	if (typeSpec->name != NULL)
-		free(typeSpec->name);
+    if (typeSpec->name != NULL)
+        free(typeSpec->name);
 
-	if (typeSpec->type == MMS_STRUCTURE) {
-		int elementCount = typeSpec->typeSpec.structure.elementCount;
-		int i;
-		for (i = 0; i < elementCount; i++) {
-			MmsVariableSpecification_destroy(typeSpec->typeSpec.structure.elements[i]);
-		}
+    if (typeSpec->type == MMS_STRUCTURE) {
+        int elementCount = typeSpec->typeSpec.structure.elementCount;
+        int i;
+        for (i = 0; i < elementCount; i++) {
+            MmsVariableSpecification_destroy(typeSpec->typeSpec.structure.elements[i]);
+        }
 
-		free(typeSpec->typeSpec.structure.elements);
-	}
-	else if (typeSpec->type == MMS_ARRAY) {
-		MmsVariableSpecification_destroy(typeSpec->typeSpec.array.elementTypeSpec);
-	}
+        free(typeSpec->typeSpec.structure.elements);
+    }
+    else if (typeSpec->type == MMS_ARRAY) {
+        MmsVariableSpecification_destroy(typeSpec->typeSpec.array.elementTypeSpec);
+    }
 
-	free(typeSpec);
+    free(typeSpec);
 }
 
-static int
+static size_t
 directChildStrLen(char* childId)
 {
-	int i = 0;
-	int childIdLen = strlen(childId);
-	while (i < childIdLen) {
-		if (*(childId + i) == '$')
-			break;
-		i++;
-	}
+    size_t i = 0;
+    size_t childIdLen = strlen(childId);
+    while (i < childIdLen) {
+        if (*(childId + i) == '$')
+            break;
+        i++;
+    }
 
-	return i;
+    return i;
 }
 
 MmsValue*
 MmsVariableSpecification_getChildValue(MmsVariableSpecification* typeSpec, MmsValue* value, char* childId)
 {
-	if (typeSpec->type == MMS_STRUCTURE) {
-		int childLen = directChildStrLen(childId);
-		int i;
-		for (i = 0; i < typeSpec->typeSpec.structure.elementCount; i++) {
+    if (typeSpec->type == MMS_STRUCTURE) {
+        size_t childLen = directChildStrLen(childId);
+        int i;
+        for (i = 0; i < typeSpec->typeSpec.structure.elementCount; i++) {
 
-			if (strlen(typeSpec->typeSpec.structure.elements[i]->name) == childLen) {
+            if (strlen(typeSpec->typeSpec.structure.elements[i]->name) == childLen) {
 
-				if (strncmp(typeSpec->typeSpec.structure.elements[i]->name, childId, childLen) == 0) {
-					if (childLen == strlen(childId)) {
-						return value->value.structure.components[i];
-					}
-					else {
-						return MmsVariableSpecification_getChildValue(typeSpec->typeSpec.structure.elements[i],
-								value->value.structure.components[i], childId + childLen + 1);
-					}
-				}
-			}
-		}
-		return NULL;
-	}
-	else
-		return NULL;
+                if (strncmp(typeSpec->typeSpec.structure.elements[i]->name, childId, childLen) == 0) {
+                    if (childLen == strlen(childId)) {
+                        return value->value.structure.components[i];
+                    }
+                    else {
+                        return MmsVariableSpecification_getChildValue(typeSpec->typeSpec.structure.elements[i],
+                                value->value.structure.components[i], childId + childLen + 1);
+                    }
+                }
+            }
+        }
+        return NULL;
+    }
+    else
+        return NULL;
 }
 
 MmsType
 MmsVariableSpecification_getType(MmsVariableSpecification* self)
 {
-	return self->type;
+    return self->type;
 }
 
 char*
 MmsVariableSpecification_getName(MmsVariableSpecification* self)
 {
-	return self->name;
+    return self->name;
 }
 
 LinkedList /* <char*> */
 MmsVariableSpecification_getStructureElements(MmsVariableSpecification* self)
 {
-	if (self->type != MMS_STRUCTURE)
-		return NULL;
+    if (self->type != MMS_STRUCTURE)
+        return NULL;
 
-	LinkedList elementNames = LinkedList_create();
+    LinkedList elementNames = LinkedList_create();
 
-	int i;
-	for (i = 0; i < self->typeSpec.structure.elementCount; i++) {
-		MmsVariableSpecification* typeSpec = self->typeSpec.structure.elements[i];
+    int i;
+    for (i = 0; i < self->typeSpec.structure.elementCount; i++) {
+        MmsVariableSpecification* typeSpec = self->typeSpec.structure.elements[i];
 
-		LinkedList_add(elementNames, copyString(typeSpec->name));
-	}
+        LinkedList_add(elementNames, copyString(typeSpec->name));
+    }
 
-	return elementNames;
+    return elementNames;
 }
-
 
 int
 MmsVariableSpecification_getSize(MmsVariableSpecification* self)
 {
-	switch(self->type) {
-	case MMS_STRUCTURE:
-		return self->typeSpec.structure.elementCount;
-	case MMS_ARRAY:
-		return self->typeSpec.array.elementCount;
-	case MMS_INTEGER:
-		return self->typeSpec.integer;
-	case MMS_UNSIGNED:
-		return self->typeSpec.unsignedInteger;
-	case MMS_FLOAT:
-		return self->typeSpec.floatingpoint.formatWidth;
-	case MMS_BIT_STRING:
-		return self->typeSpec.bitString;
-	case MMS_BINARY_TIME:
-		return self->typeSpec.binaryTime;
-	case MMS_OCTET_STRING:
-		return self->typeSpec.octetString;
-	case MMS_VISIBLE_STRING:
-		return self->typeSpec.visibleString;
-	case MMS_STRING:
-		return self->typeSpec.mmsString;
-	default:
-		return -1;
-	}
+    switch (self->type) {
+    case MMS_STRUCTURE:
+        return self->typeSpec.structure.elementCount;
+    case MMS_ARRAY:
+        return self->typeSpec.array.elementCount;
+    case MMS_INTEGER:
+        return self->typeSpec.integer;
+    case MMS_UNSIGNED:
+        return self->typeSpec.unsignedInteger;
+    case MMS_FLOAT:
+        return self->typeSpec.floatingpoint.formatWidth;
+    case MMS_BIT_STRING:
+        return self->typeSpec.bitString;
+    case MMS_BINARY_TIME:
+        return self->typeSpec.binaryTime;
+    case MMS_OCTET_STRING:
+        return self->typeSpec.octetString;
+    case MMS_VISIBLE_STRING:
+        return self->typeSpec.visibleString;
+    case MMS_STRING:
+        return self->typeSpec.mmsString;
+    default:
+        return -1;
+    }
 
 }
 
 MmsVariableSpecification*
 MmsVariableSpecification_getChildSpecificationByIndex(MmsVariableSpecification* self, int index)
 {
-	if (self->type != MMS_STRUCTURE)
-		return NULL;
+    if (self->type != MMS_STRUCTURE)
+        return NULL;
 
-	if (index >= self->typeSpec.structure.elementCount)
-		return NULL;
+    if (index >= self->typeSpec.structure.elementCount)
+        return NULL;
 
-	if (index < 0)
-		return NULL;
+    if (index < 0)
+        return NULL;
 
-	return self->typeSpec.structure.elements[index];
+    return self->typeSpec.structure.elements[index];
 }
 
 MmsVariableSpecification*
 MmsVariableSpecification_getArrayElementSpecification(MmsVariableSpecification* self)
 {
-	if (self->type != MMS_ARRAY)
-		return NULL;
+    if (self->type != MMS_ARRAY)
+        return NULL;
 
-	return self->typeSpec.array.elementTypeSpec;
+    return self->typeSpec.array.elementTypeSpec;
 }
 
 int
 MmsVariableSpecification_getExponentWidth(MmsVariableSpecification* self)
 {
-	if (self->type != MMS_FLOAT)
-		return -1;
+    if (self->type != MMS_FLOAT)
+        return -1;
 
-	return (int) self->typeSpec.floatingpoint.exponentWidth;
+    return (int) self->typeSpec.floatingpoint.exponentWidth;
 }
